+++ src/routes/AppRouter.tsx
@@ -1,5 +1,5 @@
 import { Route, Routes } from 'react-router-dom';
-import { ROUTE_PATH } from './routhPath';
+import { ROUTE_PATH } from './routePath';
 import ArchivePage from '../pages/Archive/Archive';
 import FeedbackPage from '../pages/Feedback/Feedback';
 import SubscribePage from '../pages/Subscribe/Subscribe';
